+++ src/main/java/electrodynamics/common/tile/TileElectricArcFurnace.java
@@ -173,7 +173,7 @@ public class TileElectricArcFurnace extends GenericTile implements ITickableSoun
 	
 	@Override
 	public int getComparatorSignal() {
-		return (getNumActiveProcessors() / Math.max(1, getNumProcessors())) * 15;
+		return (int) (((double) getNumActiveProcessors() / (double) Math.max(1, getNumProcessors())) * 15.0);
 	}
 
 }
