+++ src/main/java/electrodynamics/common/tile/TileLathe.java
@@ -73,7 +73,7 @@ public class TileLathe extends GenericTile implements ITickableSound {
 	
 	@Override
 	public int getComparatorSignal() {
-		return (getNumActiveProcessors() / Math.max(1, getNumProcessors())) * 15;
+		return (int) (((double) getNumActiveProcessors() / (double) Math.max(1, getNumProcessors())) * 15.0);
 	}
 
 	static {
